requirements.txt
discord.py
mss
opencv-python
pynput
requests
python-dotenv
Pillow
pystray
plyer